style: 사용되지 않는 import 제거

+++ src/main/java/org/backrow/solt/repository/BoardRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BoardRepository extends JpaRepository<Board, Long>, BoardSearch {
-    Page<Board> findByMemberMemberId(Integer memberId, Pageable pageable);
+    Page<Board> findByMemberMemberId(Long memberId, Pageable pageable);
 }
